Restructure app (#31)
