+++ src/components/types/styled.d.ts
