+++ src/markdoc/nodes/image.markdoc.ts
